feat: enhance competition router and performance
Add HTTP status checks, extract flattenEvents to utility, cache formatter, memoize sorting
